+++ src/components/fullWidthlabel.js
@@ -1,3 +1,4 @@
 import { Label } from 'reactstrap';
 
-export const FullWidthlabel = props => <Label {...props} sm="12" />;
+export const FullWidthLabelField = props => <Label {...props} sm="12" />;
+export const FullWidthLabelTextarea = props => <Label {...props} sm="5" />;
