+++ src/styles/app.module.css
@@ -3,3 +3,10 @@
   justify-content: center;
   text-align: center;
 }
+
+.app .nav {
+  border-bottom: none;
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
